change utils architecture
