+++ frontend/src/utils/rowsUtil/rowsUtil.ts
@@ -1,0 +1,13 @@
+import { v4 as uuidv4 } from "uuid";
+import { ITravel } from "../../interfaces/travels.interfaces";
+
+export const rowsUtil = (tableData: ITravel[]) => {
+    return tableData.map(({ region, country, currency }) => {
+        return {
+            id: uuidv4(),
+            region,
+            country,
+            currency,
+        };
+    });
+};
